+++ SkipIntro/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.SkipIntro")]
 [assembly: Guid("56fb522e-923d-4582-ad3d-8ec80f09237a")]
-[assembly: AssemblyVersion("1.6.0.0")]
-[assembly: AssemblyFileVersion("1.6.0.0")]
+[assembly: AssemblyVersion("1.7.0.0")]
+[assembly: AssemblyFileVersion("1.7.0.0")]